Adding a DoubleReverser implementation
This algorithm reverses the words by reversing the whole string, then
each word so they're in the correct order.

I first heard about this problem as an interview question, and it's
stuck with me because of this clever solution, which is low overhead.

I've held off on really optimizing until I have some sort of profiling.