Adding a simple profiling program for measuring optimization
This is far from a perfect profiler, but it's an easy to run and gives
some idea of the performance of the function. It uses clock time rather
than CPU time because Java has no cross-platform way of getting CPU
time. It would be nice to also measure peak memory usage, but the
garbage compiler makes it hard to get a good grasp on memory usage
while the algorithm runs.